+++ addon/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 0.1.0 beta (2026-05-12)
-- First HA Add-on release
+- First HA App release
 - Appliance questionnaire for smart plugs
 - Sniper mode: adaptive polling 20s/60s
 - Appliance program recognition